+++ src/hooks/useQuestionManagement.ts
@@ -1,0 +1,116 @@
+/**
+ * 質問管理用カスタムフック
+ * 個別質問の管理操作を提供
+ */
+
+import { useCallback } from 'react';
+import { useQuestionListStore } from '../stores/questionListStore';
+import { useErrorHandler } from './useErrorHandler';
+import type { CreateQuestionInput, UpdateQuestionInput } from '../types/data';
+
+/**
+ * 質問管理フック
+ */
+export function useQuestionManagement() {
+  const {
+    addQuestion,
+    updateQuestion,
+    deleteQuestion,
+    answerQuestion,
+    reorderQuestions,
+    sortCurrentListByAnswerStatus,
+  } = useQuestionListStore();
+
+  const { handleAsyncOperation } = useErrorHandler();
+
+  // 質問追加
+  const addQuestionToList = useCallback(
+    async (listId: string, input: CreateQuestionInput) => {
+      return await handleAsyncOperation(() => addQuestion(listId, input), {
+        loadingMessage: '質問を追加中...',
+        successMessage: '質問を追加しました',
+      });
+    },
+    [addQuestion, handleAsyncOperation]
+  );
+
+  // 質問更新
+  const updateQuestionInList = useCallback(
+    async (
+      listId: string,
+      questionId: string,
+      updates: UpdateQuestionInput
+    ) => {
+      return await handleAsyncOperation(
+        () => updateQuestion(listId, questionId, updates),
+        {
+          loadingMessage: '質問を更新中...',
+          successMessage: '質問を更新しました',
+        }
+      );
+    },
+    [updateQuestion, handleAsyncOperation]
+  );
+
+  // 質問削除
+  const deleteQuestionFromList = useCallback(
+    async (listId: string, questionId: string) => {
+      return await handleAsyncOperation(
+        () => deleteQuestion(listId, questionId),
+        {
+          loadingMessage: '質問を削除中...',
+          successMessage: '質問を削除しました',
+        }
+      );
+    },
+    [deleteQuestion, handleAsyncOperation]
+  );
+
+  // 質問回答
+  const answerQuestionInList = useCallback(
+    async (listId: string, questionId: string, answer: string) => {
+      return await handleAsyncOperation(
+        () => answerQuestion(listId, questionId, answer),
+        {
+          loadingMessage: '回答を保存中...',
+          successMessage: answer.trim()
+            ? '回答を保存しました'
+            : '回答をクリアしました',
+        }
+      );
+    },
+    [answerQuestion, handleAsyncOperation]
+  );
+
+  // 質問並び替え
+  const reorderQuestionsInList = useCallback(
+    async (listId: string, fromIndex: number, toIndex: number) => {
+      return await handleAsyncOperation(
+        () => reorderQuestions(listId, fromIndex, toIndex),
+        {
+          loadingMessage: '質問を並び替え中...',
+          successMessage: '質問を並び替えました',
+        }
+      );
+    },
+    [reorderQuestions, handleAsyncOperation]
+  );
+
+  // 回答状況でソート
+  const sortByAnswerStatus = useCallback(async () => {
+    return await handleAsyncOperation(() => sortCurrentListByAnswerStatus(), {
+      loadingMessage: '質問を並び替え中...',
+      successMessage: '未回答の質問を上部に移動しました',
+    });
+  }, [sortCurrentListByAnswerStatus, handleAsyncOperation]);
+
+  return {
+    // 質問操作
+    addQuestionToList,
+    updateQuestionInList,
+    deleteQuestionFromList,
+    answerQuestionInList,
+    reorderQuestionsInList,
+    sortByAnswerStatus,
+  };
+}
